Separate into 'events'-app
Had to comment out code in stregsystem.models which require a circular
dependency
