+++ stregsystem/views.py
@@ -22,6 +22,7 @@ from django.shortcuts import get_object_or_404, render, redirect
 from django.utils import timezone
 from django.views.decorators.csrf import csrf_exempt
 
+from events.models import TicketRecord
 from stregreport.views import fjule_party
 
 from stregsystem import parser
@@ -41,8 +42,6 @@ from stregsystem.models import (
     NamedProduct,
     ApprovalModel,
     ProductNote,
-    TicketRecord,
-    Ticket,
 )
 from stregsystem.templatetags.stregsystem_extras import money
 from stregsystem.utils import (
