+++ treo/settings.py
@@ -99,6 +99,7 @@ INSTALLED_APPS = [
     'stregreport',
     'kiosk',
     'razzia',
+    'events',
     'django_select2',
     'django.contrib.admin',
     'django.contrib.auth',
